Show add organization button on if user has valid orgs to add
